Update to v099r16, fix MSU-1 error
Updated nSide to the code base of higan v099r16, and imported a
suggested change from Cydrak that fixes an MSU-1 bug that plagues higan
v099r16.

+++ nSide/sfc-balanced/coprocessor/msu1/msu1.cpp
@@ -25,8 +25,8 @@ auto MSU1::main() -> void {
         }
       } else {
         mmio.audioPlayOffset += 4;
-        left  = (double)audioFile->readl(2) / 32768.0 * (double)mmio.audioVolume / 255.0;
-        right = (double)audioFile->readl(2) / 32768.0 * (double)mmio.audioVolume / 255.0;
+        left  = (double)(int16)audioFile->readl(2) / 32768.0 * (double)mmio.audioVolume / 255.0;
+        right = (double)(int16)audioFile->readl(2) / 32768.0 * (double)mmio.audioVolume / 255.0;
         if(dsp.mute()) left = 0, right = 0;
       }
     } else {
